Fixed Issues with Discord Enabled Checkbox
